refactor: move overlay open state inside
