+++ apps/extension/src/common/wallet-context/index.ts
@@ -1,3 +1,5 @@
 export { WalletContext } from './wallet-context'
 export { WalletProvider } from './wallet-provider'
 export { useWallet } from './use-wallet'
+export { useConnectWallet } from './use-connect-wallet'
+export { useDisconnectWallet } from './use-disconnect-wallet'
