--- apps/extension/src/contentscript/multitable-panel/constants.ts
@@ -1,1 +1,0 @@
-export const ipfs = 'https://ipfs.near.social/ipfs/'
