+++ apps/extension/src/contentscript/multitable-panel/multitable-panel.tsx
@@ -1,13 +1,7 @@
-import { EntitySourceType } from '@mweb/backend'
 import { useMutableWeb } from '@mweb/engine'
-import { useApplications } from '@mweb/react-engine'
-import { EngineProvider } from '@mweb/shared-components'
+import { EngineProvider, MiniOverlay } from '@mweb/shared-components'
 import React, { FC, useState } from 'react'
-import styled from 'styled-components'
-import { MiniOverlay } from '@mweb/shared-components'
-import { useWallet } from '../../common/wallet-context'
-import { useConnectWallet } from '../../common/wallet-context/use-connect-wallet'
-import { useDisconnectWallet } from '../../common/wallet-context/use-disconnect-wallet'
+import { useWallet, useConnectWallet, useDisconnectWallet } from '../../common/wallet-context'
 import { useSidePanel } from '../hooks/use-side-panel'
 
 export const MultitablePanel: FC = () => {
@@ -16,7 +10,6 @@ export const MultitablePanel: FC = () => {
   const { connectWallet } = useConnectWallet()
   const { disconnectWallet } = useDisconnectWallet()
   const { tree, mutations, selectedMutation, switchMutation } = useMutableWeb()
-  const [isOverlayOpen, setIsOverlayOpen] = useState(false) // ToDo: move to MiniOverlay
 
   // The notch can be opened from the extension's context menu on websites without any mutation
   if (mutations.length === 0) return null
@@ -32,7 +25,7 @@ export const MultitablePanel: FC = () => {
         selectedMutationId={selectedMutation?.id ?? null}
         onSwitchMutation={switchMutation}
       >
-        <MiniOverlay setOpen={(isOpen) => setIsOverlayOpen(isOpen)} open={isOverlayOpen} />
+        <MiniOverlay />
       </EngineProvider>
     </>
   )
