+++ libs/shared-components/src/connected-accounts/connected-account-list-item.tsx
@@ -4,7 +4,7 @@ import cn from 'classnames'
 import React, { useRef, useState } from 'react'
 import styled from 'styled-components'
 import { useOutside } from '../hooks/use-outside'
-import useCopied from '../hooks/useCopyed'
+import useCopied from '../hooks/use-copied'
 import { MoreHoriz } from './assets/icons'
 import { GithubIcon, XIcon } from './assets/resources/social'
 import { resources } from './resources'
